+++ sdk/program/src/sysvar.rs
@@ -1,0 +1,10 @@
+#[deprecated(since = "2.1.0", note = "Use `solana-sysvar-id` crate instead")]
+pub use solana_sysvar_id::{declare_deprecated_sysvar_id, declare_sysvar_id, SysvarId};
+#[allow(deprecated)]
+pub use {
+    solana_sdk_ids::sysvar::{check_id, id, ID},
+    solana_sysvar::{
+        clock, epoch_rewards, epoch_schedule, fees, instructions, is_sysvar_id, last_restart_slot,
+        recent_blockhashes, rent, slot_hashes, slot_history, stake_history, Sysvar, ALL_IDS,
+    },
+};
