+++ sdk/sysvar/Cargo.toml
@@ -1,0 +1,64 @@
+[package]
+name = "solana-sysvar"
+description = "Solana sysvar account types"
+documentation = "https://docs.rs/solana-sysvar"
+version = { workspace = true }
+authors = { workspace = true }
+repository = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+edition = { workspace = true }
+
+[dependencies]
+bincode = { workspace = true }
+bytemuck = { workspace = true }
+bytemuck_derive = { workspace = true }
+lazy_static = { workspace = true }
+qualifier_attr = { workspace = true, optional = true }
+serde = { workspace = true }
+serde_derive = { workspace = true }
+solana-account-info = { workspace = true }
+solana-clock = { workspace = true, features = ["serde", "sysvar"] }
+solana-epoch-rewards = { workspace = true, features = ["serde", "sysvar"] }
+solana-epoch-schedule = { workspace = true, features = ["serde", "sysvar"] }
+solana-fee-calculator = { workspace = true, features = ["serde"] }
+solana-hash = { workspace = true, features = ["bytemuck"] }
+solana-instruction = { workspace = true }
+solana-last-restart-slot = { workspace = true, features = ["serde", "sysvar"] }
+solana-program-entrypoint = { workspace = true }
+solana-program-error = { workspace = true }
+solana-pubkey = { workspace = true }
+solana-rent = { workspace = true, features = ["serde", "sysvar"] }
+solana-sanitize = { workspace = true }
+solana-sdk-ids = { workspace = true }
+solana-sdk-macro = { workspace = true }
+solana-serialize-utils = { workspace = true }
+solana-slot-hashes = { workspace = true, features = ["serde", "sysvar"] }
+solana-slot-history = { workspace = true, features = ["serde", "sysvar"] }
+solana-sysvar-id = { workspace = true }
+
+[target.'cfg(not(target_os = "solana"))'.dependencies]
+base64 = { workspace = true }
+bitflags = { workspace = true }
+solana-program-memory = { workspace = true }
+
+[dev-dependencies]
+anyhow = { workspace = true }
+serial_test = { workspace = true }
+solana-account = { workspace = true }
+solana-msg = { workspace = true }
+solana-program = { path = "../program" }
+solana-sdk = { path = ".." }
+solana-sha256-hasher = { workspace = true }
+test-case = { workspace = true }
+
+[features]
+dev-context-only-utils = ["dep:qualifier_attr"]
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+all-features = true
+rustdoc-args = ["--cfg=docsrs"]
+
+[lints]
+workspace = true
