+++ sdk/sysvar/src/instructions.rs
@@ -14,7 +14,7 @@
 //! instruction sysvar is accessed through several free functions within this
 //! module.
 //!
-//! [`Sysvar`]: crate::sysvar::Sysvar
+//! [`Sysvar`]: crate::Sysvar
 //!
 //! See also the Solana [documentation on the instructions sysvar][sdoc].
 //!
@@ -54,7 +54,7 @@ use {
 /// but does not implement the [`Sysvar`] trait.
 ///
 /// [`SysvarId`]: https://docs.rs/solana-sysvar-id/latest/solana_sysvar_id/trait.SysvarId.html
-/// [`Sysvar`]: crate::sysvar::Sysvar
+/// [`Sysvar`]: crate::Sysvar
 ///
 /// Use the free functions in this module to access the instructions sysvar.
 pub struct Instructions();
@@ -95,6 +95,7 @@ bitflags! {
 //   67..69 - data len - u16
 //   69..data_len - data
 #[cfg(not(target_os = "solana"))]
+#[cfg_attr(feature = "dev-context-only-utils", qualifiers(pub))]
 fn serialize_instructions(instructions: &[BorrowedInstruction]) -> Vec<u8> {
     // 64 bytes is a reasonable guess, calculating exactly is slower in benchmarks
     let mut data = Vec::with_capacity(instructions.len() * (32 * 2));
@@ -165,6 +166,7 @@ pub fn store_current_index(data: &mut [u8], instruction_index: u16) {
     data[last_index..last_index + 2].copy_from_slice(&instruction_index.to_le_bytes());
 }
 
+#[cfg_attr(feature = "dev-context-only-utils", qualifiers(pub))]
 fn deserialize_instruction(index: usize, data: &[u8]) -> Result<Instruction, SanitizeError> {
     const IS_SIGNER_BIT: usize = 0;
     const IS_WRITABLE_BIT: usize = 1;
@@ -274,17 +276,7 @@ pub fn get_instruction_relative(
 
 #[cfg(test)]
 mod tests {
-    use {
-        super::*,
-        solana_instruction::AccountMeta,
-        solana_program::message::{Message as LegacyMessage, SanitizedMessage},
-        solana_pubkey::Pubkey,
-        std::collections::HashSet,
-    };
-
-    fn new_sanitized_message(message: LegacyMessage) -> SanitizedMessage {
-        SanitizedMessage::try_from_legacy_message(message, &HashSet::default()).unwrap()
-    }
+    use super::*;
 
     #[test]
     fn test_load_store_instruction() {
@@ -294,257 +286,5 @@ mod tests {
         let index = load_current_index(&data);
         assert_eq!(index, 3);
         assert_eq!([4u8; 8], data[0..8]);
-    }
-
-    #[test]
-    fn test_load_instruction_at_checked() {
-        let instruction0 = Instruction::new_with_bincode(
-            Pubkey::new_unique(),
-            &0,
-            vec![AccountMeta::new(Pubkey::new_unique(), false)],
-        );
-        let instruction1 = Instruction::new_with_bincode(
-            Pubkey::new_unique(),
-            &0,
-            vec![AccountMeta::new(Pubkey::new_unique(), false)],
-        );
-        let message = LegacyMessage::new(
-            &[instruction0.clone(), instruction1.clone()],
-            Some(&Pubkey::new_unique()),
-        );
-        let sanitized_message = new_sanitized_message(message);
-
-        let key = id();
-        let mut lamports = 0;
-        let mut data = construct_instructions_data(&sanitized_message.decompile_instructions());
-        let owner = solana_sdk_ids::sysvar::id();
-        let mut account_info = AccountInfo::new(
-            &key,
-            false,
-            false,
-            &mut lamports,
-            &mut data,
-            &owner,
-            false,
-            0,
-        );
-
-        assert_eq!(
-            instruction0,
-            load_instruction_at_checked(0, &account_info).unwrap()
-        );
-        assert_eq!(
-            instruction1,
-            load_instruction_at_checked(1, &account_info).unwrap()
-        );
-        assert_eq!(
-            Err(ProgramError::InvalidArgument),
-            load_instruction_at_checked(2, &account_info)
-        );
-
-        let key = Pubkey::new_unique();
-        account_info.key = &key;
-        assert_eq!(
-            Err(ProgramError::UnsupportedSysvar),
-            load_instruction_at_checked(2, &account_info)
-        );
-    }
-
-    #[test]
-    fn test_load_current_index_checked() {
-        let instruction0 = Instruction::new_with_bincode(
-            Pubkey::new_unique(),
-            &0,
-            vec![AccountMeta::new(Pubkey::new_unique(), false)],
-        );
-        let instruction1 = Instruction::new_with_bincode(
-            Pubkey::new_unique(),
-            &0,
-            vec![AccountMeta::new(Pubkey::new_unique(), false)],
-        );
-        let message =
-            LegacyMessage::new(&[instruction0, instruction1], Some(&Pubkey::new_unique()));
-        let sanitized_message = new_sanitized_message(message);
-
-        let key = id();
-        let mut lamports = 0;
-        let mut data = construct_instructions_data(&sanitized_message.decompile_instructions());
-        store_current_index(&mut data, 1);
-        let owner = solana_sdk_ids::sysvar::id();
-        let mut account_info = AccountInfo::new(
-            &key,
-            false,
-            false,
-            &mut lamports,
-            &mut data,
-            &owner,
-            false,
-            0,
-        );
-
-        assert_eq!(1, load_current_index_checked(&account_info).unwrap());
-        {
-            let mut data = account_info.try_borrow_mut_data().unwrap();
-            store_current_index(&mut data, 0);
-        }
-        assert_eq!(0, load_current_index_checked(&account_info).unwrap());
-
-        let key = Pubkey::new_unique();
-        account_info.key = &key;
-        assert_eq!(
-            Err(ProgramError::UnsupportedSysvar),
-            load_current_index_checked(&account_info)
-        );
-    }
-
-    #[test]
-    fn test_get_instruction_relative() {
-        let instruction0 = Instruction::new_with_bincode(
-            Pubkey::new_unique(),
-            &0,
-            vec![AccountMeta::new(Pubkey::new_unique(), false)],
-        );
-        let instruction1 = Instruction::new_with_bincode(
-            Pubkey::new_unique(),
-            &0,
-            vec![AccountMeta::new(Pubkey::new_unique(), false)],
-        );
-        let instruction2 = Instruction::new_with_bincode(
-            Pubkey::new_unique(),
-            &0,
-            vec![AccountMeta::new(Pubkey::new_unique(), false)],
-        );
-        let message = LegacyMessage::new(
-            &[
-                instruction0.clone(),
-                instruction1.clone(),
-                instruction2.clone(),
-            ],
-            Some(&Pubkey::new_unique()),
-        );
-        let sanitized_message = new_sanitized_message(message);
-
-        let key = id();
-        let mut lamports = 0;
-        let mut data = construct_instructions_data(&sanitized_message.decompile_instructions());
-        store_current_index(&mut data, 1);
-        let owner = solana_sdk_ids::sysvar::id();
-        let mut account_info = AccountInfo::new(
-            &key,
-            false,
-            false,
-            &mut lamports,
-            &mut data,
-            &owner,
-            false,
-            0,
-        );
-
-        assert_eq!(
-            Err(ProgramError::InvalidArgument),
-            get_instruction_relative(-2, &account_info)
-        );
-        assert_eq!(
-            instruction0,
-            get_instruction_relative(-1, &account_info).unwrap()
-        );
-        assert_eq!(
-            instruction1,
-            get_instruction_relative(0, &account_info).unwrap()
-        );
-        assert_eq!(
-            instruction2,
-            get_instruction_relative(1, &account_info).unwrap()
-        );
-        assert_eq!(
-            Err(ProgramError::InvalidArgument),
-            get_instruction_relative(2, &account_info)
-        );
-        {
-            let mut data = account_info.try_borrow_mut_data().unwrap();
-            store_current_index(&mut data, 0);
-        }
-        assert_eq!(
-            Err(ProgramError::InvalidArgument),
-            get_instruction_relative(-1, &account_info)
-        );
-        assert_eq!(
-            instruction0,
-            get_instruction_relative(0, &account_info).unwrap()
-        );
-        assert_eq!(
-            instruction1,
-            get_instruction_relative(1, &account_info).unwrap()
-        );
-        assert_eq!(
-            instruction2,
-            get_instruction_relative(2, &account_info).unwrap()
-        );
-        assert_eq!(
-            Err(ProgramError::InvalidArgument),
-            get_instruction_relative(3, &account_info)
-        );
-
-        let key = Pubkey::new_unique();
-        account_info.key = &key;
-        assert_eq!(
-            Err(ProgramError::UnsupportedSysvar),
-            get_instruction_relative(0, &account_info)
-        );
-    }
-
-    #[test]
-    fn test_serialize_instructions() {
-        let program_id0 = Pubkey::new_unique();
-        let program_id1 = Pubkey::new_unique();
-        let id0 = Pubkey::new_unique();
-        let id1 = Pubkey::new_unique();
-        let id2 = Pubkey::new_unique();
-        let id3 = Pubkey::new_unique();
-        let instructions = vec![
-            Instruction::new_with_bincode(program_id0, &0, vec![AccountMeta::new(id0, false)]),
-            Instruction::new_with_bincode(program_id0, &0, vec![AccountMeta::new(id1, true)]),
-            Instruction::new_with_bincode(
-                program_id1,
-                &0,
-                vec![AccountMeta::new_readonly(id2, false)],
-            ),
-            Instruction::new_with_bincode(
-                program_id1,
-                &0,
-                vec![AccountMeta::new_readonly(id3, true)],
-            ),
-        ];
-
-        let message = LegacyMessage::new(&instructions, Some(&id1));
-        let sanitized_message = new_sanitized_message(message);
-        let serialized = serialize_instructions(&sanitized_message.decompile_instructions());
-
-        // assert that deserialize_instruction is compatible with SanitizedMessage::serialize_instructions
-        for (i, instruction) in instructions.iter().enumerate() {
-            assert_eq!(
-                deserialize_instruction(i, &serialized).unwrap(),
-                *instruction
-            );
-        }
-    }
-
-    #[test]
-    fn test_decompile_instructions_out_of_bounds() {
-        let program_id0 = Pubkey::new_unique();
-        let id0 = Pubkey::new_unique();
-        let id1 = Pubkey::new_unique();
-        let instructions = vec![
-            Instruction::new_with_bincode(program_id0, &0, vec![AccountMeta::new(id0, false)]),
-            Instruction::new_with_bincode(program_id0, &0, vec![AccountMeta::new(id1, true)]),
-        ];
-
-        let message = LegacyMessage::new(&instructions, Some(&id1));
-        let sanitized_message = new_sanitized_message(message);
-        let serialized = serialize_instructions(&sanitized_message.decompile_instructions());
-        assert_eq!(
-            deserialize_instruction(instructions.len(), &serialized).unwrap_err(),
-            SanitizeError::IndexOutOfBounds,
-        );
     }
 }
