+++ sdk/sysvar/src/program_stubs.rs
@@ -113,25 +113,25 @@ pub trait SyscallStubs: Sync + Send {
 struct DefaultSyscallStubs {}
 impl SyscallStubs for DefaultSyscallStubs {}
 
-pub(crate) fn sol_log(message: &str) {
+pub fn sol_log(message: &str) {
     SYSCALL_STUBS.read().unwrap().sol_log(message);
 }
 
-pub(crate) fn sol_log_64(arg1: u64, arg2: u64, arg3: u64, arg4: u64, arg5: u64) {
+pub fn sol_log_64(arg1: u64, arg2: u64, arg3: u64, arg4: u64, arg5: u64) {
     sol_log(&format!(
         "{arg1:#x}, {arg2:#x}, {arg3:#x}, {arg4:#x}, {arg5:#x}"
     ));
 }
 
-pub(crate) fn sol_log_compute_units() {
+pub fn sol_log_compute_units() {
     SYSCALL_STUBS.read().unwrap().sol_log_compute_units();
 }
 
-pub(crate) fn sol_remaining_compute_units() -> u64 {
+pub fn sol_remaining_compute_units() -> u64 {
     SYSCALL_STUBS.read().unwrap().sol_remaining_compute_units()
 }
 
-pub(crate) fn sol_invoke_signed(
+pub fn sol_invoke_signed(
     instruction: &Instruction,
     account_infos: &[AccountInfo],
     signers_seeds: &[&[&[u8]]],
@@ -181,33 +181,33 @@ pub(crate) fn sol_get_last_restart_slot(var_addr: *mut u8) -> u64 {
         .sol_get_last_restart_slot(var_addr)
 }
 
-pub(crate) fn sol_get_epoch_stake(vote_address: *const u8) -> u64 {
+pub fn sol_get_epoch_stake(vote_address: *const u8) -> u64 {
     SYSCALL_STUBS
         .read()
         .unwrap()
         .sol_get_epoch_stake(vote_address)
 }
 
-pub(crate) fn sol_get_return_data() -> Option<(Pubkey, Vec<u8>)> {
+pub fn sol_get_return_data() -> Option<(Pubkey, Vec<u8>)> {
     SYSCALL_STUBS.read().unwrap().sol_get_return_data()
 }
 
-pub(crate) fn sol_set_return_data(data: &[u8]) {
+pub fn sol_set_return_data(data: &[u8]) {
     SYSCALL_STUBS.read().unwrap().sol_set_return_data(data)
 }
 
-pub(crate) fn sol_log_data(data: &[&[u8]]) {
+pub fn sol_log_data(data: &[&[u8]]) {
     SYSCALL_STUBS.read().unwrap().sol_log_data(data)
 }
 
-pub(crate) fn sol_get_processed_sibling_instruction(index: usize) -> Option<Instruction> {
+pub fn sol_get_processed_sibling_instruction(index: usize) -> Option<Instruction> {
     SYSCALL_STUBS
         .read()
         .unwrap()
         .sol_get_processed_sibling_instruction(index)
 }
 
-pub(crate) fn sol_get_stack_height() -> u64 {
+pub fn sol_get_stack_height() -> u64 {
     SYSCALL_STUBS.read().unwrap().sol_get_stack_height()
 }
