+++ sdk/sysvar/src/rent.rs
@@ -19,7 +19,7 @@
 //! ```no_run
 //! # use solana_account_info::AccountInfo;
 //! # use solana_msg::msg;
-//! # use solana_program::sysvar::Sysvar;
+//! # use solana_sysvar::Sysvar;
 //! # use solana_program_error::{ProgramError, ProgramResult};
 //! # use solana_pubkey::Pubkey;
 //! # use solana_rent::Rent;
@@ -55,7 +55,7 @@
 //! ```
 //! # use solana_account_info::{AccountInfo, next_account_info};
 //! # use solana_msg::msg;
-//! # use solana_program::sysvar::Sysvar;
+//! # use solana_sysvar::Sysvar;
 //! # use solana_program_error::{ProgramError, ProgramResult};
 //! # use solana_pubkey::Pubkey;
 //! # use solana_rent::Rent;
@@ -123,7 +123,7 @@
 //! # Ok::<(), anyhow::Error>(())
 //! ```
 use {
-    crate::{impl_sysvar_get, sysvar::Sysvar},
+    crate::{impl_sysvar_get, Sysvar},
     solana_program_error::ProgramError,
 };
 pub use {
