+++ sdk/sysvar/src/slot_hashes.rs
@@ -21,14 +21,15 @@
 //! # use solana_program::example_mocks::solana_rpc_client;
 //! # use solana_sdk::account::Account;
 //! # use solana_rpc_client::rpc_client::RpcClient;
-//! # use solana_sdk::sysvar::slot_hashes::{self, SlotHashes};
+//! # use solana_sdk_ids::sysvar::slot_hashes;
+//! # use solana_slot_hashes::SlotHashes;
 //! # use anyhow::Result;
 //! #
 //! fn print_sysvar_slot_hashes(client: &RpcClient) -> Result<()> {
 //! #   client.set_get_account_response(slot_hashes::ID, Account {
 //! #       lamports: 1009200,
 //! #       data: vec![1, 0, 0, 0, 0, 0, 0, 0, 86, 190, 235, 7, 0, 0, 0, 0, 133, 242, 94, 158, 223, 253, 207, 184, 227, 194, 235, 27, 176, 98, 73, 3, 175, 201, 224, 111, 21, 65, 73, 27, 137, 73, 229, 19, 255, 192, 193, 126],
-//! #       owner: solana_sdk::system_program::ID,
+//! #       owner: solana_sdk_ids::system_program::ID,
 //! #       executable: false,
 //! #       rent_epoch: 307,
 //! # });
@@ -46,7 +47,7 @@
 //! ```
 
 use {
-    crate::sysvar::{get_sysvar, Sysvar},
+    crate::{get_sysvar, Sysvar},
     bytemuck_derive::{Pod, Zeroable},
     solana_account_info::AccountInfo,
     solana_clock::Slot,
@@ -217,9 +218,8 @@ fn get_pod_slot_hashes() -> Result<Vec<PodSlotHash>, ProgramError> {
 #[cfg(test)]
 mod tests {
     use {
-        super::*, crate::sysvar::tests::mock_get_sysvar_syscall, serial_test::serial,
-        solana_hash::Hash, solana_sha256_hasher::hash, solana_slot_hashes::MAX_ENTRIES,
-        test_case::test_case,
+        super::*, crate::tests::mock_get_sysvar_syscall, serial_test::serial, solana_hash::Hash,
+        solana_sha256_hasher::hash, solana_slot_hashes::MAX_ENTRIES, test_case::test_case,
     };
 
     #[test]
